make sure NDEBUG is not defined
The NDEBUG macro is defined by default for all pico builds except
DEBUG. Defining NDEBUG makes the assert() macro generate no code.
Each function call that adds a CLI command is wrapped in an assert
macro to make sure it did not fail, usually because there is no
space for it in the array that holds the CLI commands.